reimplementing the whole SQLDataFrame.
With new slots, and more efficient strategies in matching between the
@tblData and @keyData when extracting certerin rows and columns. Slots
of @dbkey saves the user-chosen columns names to uniquely identify each
row of data. @partitionID is optionally user-chosen column for very large
data such as BigQuery tables, so that the row ids could be generated based
on sorting within a specific categorical variable. @pidRle is an Rle object
when @partitionID is not NULL, and used to save the orders of the whole table.
@dim and @dimnames are updated everytime after subsetting or operations
such as filter, select, mutate, filter etc. to avoid unnecessary expensive
queries, especially for BigQuery tables. @ridx is used to indicate the row index
matching to the sorted @tblData after data operations, so that a randomly sampled
or duplicate rows could be returned.

After the reimplementation, the major @tblData and @keyData only have lazy
tibbles, and all (or most?) operations are based on lazy operations between
these two slots. Key columns are not realized in memory as before reimplementation
(might be a different case for MySQL, will find out when reach to that point). With
the newly added @partitionID and @pidRle slots, very large BigQuery tables are expected
to be able to represented using SQLDataFrame. Currently we are using 8.4GB data for testing
(bigquery-public-data/human_variant_annotation/ensembl_1000genomes_phase3_hg38_release93),
and it seems to be working nicely.

Will continue working on the functions to resume all previous functionalities, and
after that, test on SQLite and MySQL, especially for those across-database cases.